+++ src/shared/components/create-case-filters/create-case-filters.component.spec.ts
@@ -1,23 +1,14 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { Component, DebugElement, EventEmitter, Input, Output } from '@angular/core';
+import { DebugElement } from '@angular/core';
 import { By } from '@angular/platform-browser';
 import { FormGroup, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { CreateCaseFiltersComponent } from './create-case-filters.component';
-import { Router } from '@angular/router';
-import { Subject } from 'rxjs/Subject';
 import createSpyObj = jasmine.createSpyObj;
-import { Observable } from 'rxjs';
 import { CaseTypeLite } from '../../domain/definition/case-type-lite.model';
 import { Jurisdiction } from '../../domain/definition/jurisdiction.model';
 import { CaseEvent } from '../../domain/definition/case-event.model';
-import { JurisdictionService } from '../../services/jurisdiction/jurisdiction.service';
 import { OrderService } from '../../services/order/order.service';
 import { AlertService } from '../../services/alert/alert.service';
-import { DefinitionsService } from '../../services/definitions/definitions.service';
-import { text, attr } from '../../test/helpers';
-import { CallbackErrorsContext } from '../error/domain/error-context';
-import { CaseViewerComponent } from '../case-viewer/case-viewer.component';
-import { HttpError } from '../../domain/http/http-error.model';
 
 const EVENT_ID_1 = 'ID_1';
 const EVENT_NAME_1 = 'Event one';
@@ -225,31 +216,21 @@ const CASE_EVENTS_NO_PRE_STATES: CaseEvent[] = [
 
 const SORTED_CASE_EVENTS: CaseEvent[] = [...CASE_EVENTS_NO_PRE_STATES];
 
-@Component({
-  selector: 'ccd-callback-errors',
-  template: ``
-})
-class CallbackErrorsComponent {
-
-  @Input()
-  triggerTextIgnore: string;
-  @Input()
-  triggerTextContinue: string;
-  @Input()
-  callbackErrorsSubject: Subject<any> = new Subject();
-  @Output()
-  callbackErrorsContext: EventEmitter<any> = new EventEmitter();
-
-}
-
-let mockDefinitionsService;
-let mockRouter: any;
 let mockOrderService: any;
-let mockCallbackErrorSubject: any;
 let mockAlertService: any;
-let jurisdictionService: JurisdictionService;
 
 const TEST_FORM_GROUP = new FormGroup({});
+
+const changeDummy = (jurisdictions) => {
+  return {
+    jurisdictions: {
+      isFirstChange: () => true,
+      previousValue: null,
+      firstChange: true,
+      currentValue: jurisdictions
+    }
+  };
+}
 
 describe('CreateCaseFiltersComponent', () => {
 
@@ -261,21 +242,11 @@ describe('CreateCaseFiltersComponent', () => {
   const $SELECT_CASE_TYPE = By.css('#cc-case-type');
   const $SELECT_EVENT = By.css('#cc-event');
   const $SELECT_BUTTON = By.css('button');
-  const $ERROR_SUMMARY = By.css('.error-summary');
-  const $ERROR_MESSAGE = By.css('p');
-  const $ERROR_FIELD_MESSAGES = By.css('ul');
-  const $SUBMIT_BUTTON = By.css('form button[type=submit]');
 
   beforeEach(async(() => {
     mockOrderService = createSpyObj<OrderService>('orderService', ['sort']);
     mockOrderService.sort.and.returnValue(SORTED_CASE_EVENTS);
-    mockDefinitionsService = createSpyObj('mockDefinitionsService', ['getJurisdictions']);
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_2]));
-    mockRouter = createSpyObj<Router>('router', ['navigate']);
-    mockRouter.navigate.and.returnValue(Promise.resolve(true));
-    mockCallbackErrorSubject = createSpyObj<any>('callbackErrorSubject', ['next']);
     mockAlertService = createSpyObj<AlertService>('alertService', ['clear']);
-    jurisdictionService = new JurisdictionService();
 
     TestBed
       .configureTestingModule({
@@ -284,59 +255,49 @@ describe('CreateCaseFiltersComponent', () => {
           ReactiveFormsModule
         ],
         declarations: [
-          CreateCaseFiltersComponent,
-          CallbackErrorsComponent
+          CreateCaseFiltersComponent
         ], providers: [
-          { provide: Router, useValue: mockRouter },
           { provide: OrderService, useValue: mockOrderService },
-          { provide: AlertService, useValue: mockAlertService },
-          { provide: JurisdictionService, useValue: jurisdictionService },
-          { provide: DefinitionsService, useValue: mockDefinitionsService }
+          { provide: AlertService, useValue: mockAlertService }
         ]
       })
       .compileComponents();
 
     fixture = TestBed.createComponent(CreateCaseFiltersComponent);
     component = fixture.componentInstance;
+    component.jurisdictions = [];
     component.formGroup = TEST_FORM_GROUP;
-    component.callbackErrorsSubject = mockCallbackErrorSubject;
 
     de = fixture.debugElement;
     fixture.detectChanges();
+    component.initControls();
   }));
 
   it('should select the jurisdiction if there is only one jurisdiction', () => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_1]));
-    fixture.detectChanges();
-    component.ngOnInit();
-    fixture.detectChanges();
+    component.jurisdictions = [JURISDICTION_1];
+    component.ngOnChanges(changeDummy([JURISDICTION_1]));
     expect(component.filterJurisdictionControl.value).toBe(JURISDICTION_1.id);
   });
 
   it('should select the caseType if there is only one caseType', () => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_1]));
-    fixture.detectChanges();
-    component.ngOnInit();
-    fixture.detectChanges();
+    component.jurisdictions = [JURISDICTION_1];
+    component.ngOnChanges(changeDummy([JURISDICTION_1]));
     expect(component.filterJurisdictionControl.value).toBe(JURISDICTION_1.id);
     expect(component.filterCaseTypeControl.value).toBe('CT0');
   });
 
   it('should select the event if there is only one event', () => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_SINGLE_EVENT]));
+    component.jurisdictions = [JURISDICTION_SINGLE_EVENT];
     mockOrderService.sort.and.returnValue(SINGLE_EVENT);
-    fixture.detectChanges();
-    component.ngOnInit();
-    fixture.detectChanges();
+    component.ngOnChanges(changeDummy([JURISDICTION_SINGLE_EVENT]));
     expect(component.filterJurisdictionControl.value).toBe(JURISDICTION_SINGLE_EVENT.id);
     expect(component.filterCaseTypeControl.value).toBe('CT0');
     expect(component.filterEventControl.value).toBe(EVENT_ID_1);
   });
 
   it('should sort events', () => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_1]));
-    fixture.detectChanges();
-    component.ngOnInit();
+    component.jurisdictions = [JURISDICTION_1];
+    component.ngOnChanges(changeDummy([JURISDICTION_1]));
     component.filterJurisdictionControl.setValue(JURISDICTION_1.id);
     component.onJurisdictionIdChange();
     component.filterCaseTypeControl.setValue(CASE_TYPES_1[0].id);
@@ -348,9 +309,8 @@ describe('CreateCaseFiltersComponent', () => {
   });
 
   it('should initialise jurisdiction selector with given jurisdictions and no selection', () => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_1, JURISDICTION_2]));
-    fixture.detectChanges();
-    component.ngOnInit();
+    component.jurisdictions = [JURISDICTION_1, JURISDICTION_2];
+    component.ngOnChanges((changeDummy([JURISDICTION_1, JURISDICTION_2])));
     component.filterJurisdictionControl.setValue('');
     component.onJurisdictionIdChange();
     fixture.detectChanges();
@@ -372,9 +332,8 @@ describe('CreateCaseFiltersComponent', () => {
   });
 
   it('should update selected jurisdiction', async(() => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_SINGLE_EVENT]));
-    fixture.detectChanges();
-    component.ngOnInit();
+    component.jurisdictions = [JURISDICTION_SINGLE_EVENT];
+    component.ngOnChanges(changeDummy([JURISDICTION_SINGLE_EVENT]));
     component.filterJurisdictionControl.setValue(JURISDICTION_SINGLE_EVENT.id);
     component.onJurisdictionIdChange();
     fixture.detectChanges();
@@ -386,9 +345,8 @@ describe('CreateCaseFiltersComponent', () => {
   }));
 
   it('should initialise case type selector with types from selected jurisdiction but no events', async(() => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_2]));
-    fixture.detectChanges();
-    component.ngOnInit();
+    component.jurisdictions = [JURISDICTION_2];
+    component.ngOnChanges(changeDummy([JURISDICTION_2]));
     component.filterJurisdictionControl.setValue(JURISDICTION_2.id);
     component.onJurisdictionIdChange();
     fixture.detectChanges();
@@ -414,9 +372,8 @@ describe('CreateCaseFiltersComponent', () => {
   }));
 
   it('should update selected case type', async(() => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_2]));
-    fixture.detectChanges();
-    component.ngOnInit();
+    component.jurisdictions = [JURISDICTION_2];
+    component.ngOnChanges(changeDummy([JURISDICTION_2]));
     component.filterJurisdictionControl.setValue(JURISDICTION_2.id);
     component.onJurisdictionIdChange();
     component.filterCaseTypeControl.setValue(CASE_TYPES_2[2].id);
@@ -428,9 +385,8 @@ describe('CreateCaseFiltersComponent', () => {
   }));
 
   it('should disable case type and event if jurisdiction not selected', async(() => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_1, JURISDICTION_2]));
-    fixture.detectChanges();
-    component.ngOnInit();
+    component.jurisdictions = [JURISDICTION_1, JURISDICTION_2];
+    component.ngOnChanges(changeDummy([JURISDICTION_1, JURISDICTION_2]));
     component.filterJurisdictionControl.setValue('');
     component.onJurisdictionIdChange();
     fixture.detectChanges();
@@ -456,9 +412,8 @@ describe('CreateCaseFiltersComponent', () => {
   }));
 
   it('should initialise event selector from case type with no pre states', async(() => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_2]));
-    fixture.detectChanges();
-    component.ngOnInit();
+    component.jurisdictions = [JURISDICTION_2];
+    component.ngOnChanges(changeDummy([JURISDICTION_2]));
     component.filterJurisdictionControl.setValue(JURISDICTION_2.id);
     component.onJurisdictionIdChange();
     component.filterCaseTypeControl.setValue(CASE_TYPES_2[2].id);
@@ -496,9 +451,8 @@ describe('CreateCaseFiltersComponent', () => {
   }));
 
   it('should reset case type back to empty disabled if set before and jurisdiction changed to empty', async(() => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_2]));
-    fixture.detectChanges();
-    component.ngOnInit();
+    component.jurisdictions = [JURISDICTION_2];
+    component.ngOnChanges(changeDummy([JURISDICTION_2]));
     component.filterJurisdictionControl.setValue(JURISDICTION_2.id);
     component.onJurisdictionIdChange();
     fixture.detectChanges();
@@ -523,9 +477,8 @@ describe('CreateCaseFiltersComponent', () => {
   }));
 
   it('should reset event back to default if set before and case type changed', async(() => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_2]));
-    fixture.detectChanges();
-    component.ngOnInit();
+    component.jurisdictions = [JURISDICTION_2];
+    component.ngOnChanges(changeDummy([JURISDICTION_2]));
     component.filterJurisdictionControl.setValue(JURISDICTION_2.id);
     component.onJurisdictionIdChange();
     component.filterCaseTypeControl.setValue(CASE_TYPES_2[0].id);
@@ -557,9 +510,8 @@ describe('CreateCaseFiltersComponent', () => {
   }));
 
   it('should reset event back to default if set before and jurisdiction changed', async(() => {
-    mockDefinitionsService.getJurisdictions.and.returnValue(Observable.of([JURISDICTION_2]));
-    fixture.detectChanges();
-    component.ngOnInit();
+    component.jurisdictions = [JURISDICTION_2];
+    component.ngOnChanges(changeDummy([JURISDICTION_2]));
     component.filterJurisdictionControl.setValue(JURISDICTION_2.id);
     component.onJurisdictionIdChange();
     component.filterCaseTypeControl.setValue(CASE_TYPES_2[0].id);
@@ -618,7 +570,7 @@ describe('CreateCaseFiltersComponent', () => {
     expect(button.nativeElement.disabled).toBeFalsy();
   }));
 
-  it('should load create case page when form fields selected and Go button clicked', async(() => {
+  it('should return selected object when form fields selected and Go button clicked', async(() => {
     component.selected.jurisdiction = JURISDICTION_2;
     component.selected.caseType = CASE_TYPES_2[2];
     component.selectedCaseTypeEvents = CASE_TYPE.events;
@@ -626,302 +578,18 @@ describe('CreateCaseFiltersComponent', () => {
     component.filterEventControl.setValue(EVENT_ID_2);
     component.onEventIdChange();
 
+    spyOn(component.selectionSubmitted, 'emit');
+
     fixture.detectChanges();
     let button = de.query($SELECT_BUTTON);
     button.nativeElement.click();
 
-    expect(mockRouter.navigate).toHaveBeenCalledWith(
-      ['/create/case', JURISDICTION_2.id, CASE_TYPES_2[2].id, EVENT_ID_2],
-      { queryParams: {}});
+    expect(component.selectionSubmitted.emit).toHaveBeenCalledWith({
+      jurisdictionId: JURISDICTION_2.id,
+      caseTypeId: CASE_TYPES_2[2].id,
+      eventId: EVENT_ID_2
+    });
 
   }));
 
-  it('should notify user about errors/warnings when fields selected and button clicked and response with callback errors/warnings', () => {
-    component.selected.jurisdiction = JURISDICTION_2;
-    component.selected.caseType = CASE_TYPES_2[2];
-    component.selectedCaseTypeEvents = CASE_TYPE.events;
-    component.selected.event = CASE_TYPE.events[1];
-    component.filterEventControl.setValue(EVENT_ID_2);
-    component.onEventIdChange();
-    const VALID_ERROR = {
-      callbackErrors: ['error1', 'error2'],
-      callbackWarnings: ['warning1', 'warning2']
-    };
-    mockRouter.navigate.and.returnValue({ catch : (error) => error(VALID_ERROR)});
-
-    fixture.detectChanges();
-    let button = de.query($SELECT_BUTTON);
-    button.nativeElement.click();
-    fixture.detectChanges();
-
-    expect(mockRouter.navigate).toHaveBeenCalledWith(
-      ['/create/case', JURISDICTION_2.id, CASE_TYPES_2[2].id, EVENT_ID_2],
-      { queryParams: {}});
-    expect(mockCallbackErrorSubject.next).toHaveBeenCalledWith(VALID_ERROR);
-  });
-
-  it('should notify user about field validation errors when response with field validation errors', () => {
-    component.selected.jurisdiction = JURISDICTION_2;
-    component.selected.caseType = CASE_TYPES_2[2];
-    component.selectedCaseTypeEvents = CASE_TYPE.events;
-    component.selected.event = CASE_TYPE.events[1];
-    component.filterEventControl.setValue(EVENT_ID_2);
-    component.onEventIdChange();
-    const FIELD_ERRORS = [
-      {
-        message: 'This field1 failed validation'
-      },
-      {
-        message: 'This field2 failed validation'
-      }
-    ];
-    const VALID_ERROR = {
-      message: 'Field validation failed',
-      details: {
-        field_errors: FIELD_ERRORS
-      }
-    };
-    mockRouter.navigate.and.returnValue({ catch : (error) => error(VALID_ERROR)});
-
-    fixture.detectChanges();
-    let button = de.query($SELECT_BUTTON);
-    button.nativeElement.click();
-    fixture.detectChanges();
-
-    expect(mockRouter.navigate).toHaveBeenCalledWith(
-      ['/create/case', JURISDICTION_2.id, CASE_TYPES_2[2].id, EVENT_ID_2],
-      { queryParams: {}});
-    expect(mockCallbackErrorSubject.next).toHaveBeenCalledWith(VALID_ERROR);
-
-    let errorElement = de.query($ERROR_SUMMARY);
-    expect(errorElement).toBeTruthy();
-    let errorMessage = errorElement.query($ERROR_MESSAGE);
-    expect(text(errorMessage)).toBe('Field validation failed');
-    let errorFieldMessages = errorElement.query($ERROR_FIELD_MESSAGES);
-    expect(errorFieldMessages.children.length).toBe(2);
-    expect(errorFieldMessages.children[0].nativeElement.textContent).toContain('This field1 failed validation');
-    expect(errorFieldMessages.children[1].nativeElement.textContent).toContain('This field2 failed validation');
-  });
-
-  it('should remove field validation errors when errors previously but new jurisdiction selected', () => {
-    component.selected.jurisdiction = JURISDICTION_2;
-    component.selected.caseType = CASE_TYPES_2[2];
-    component.selectedCaseTypeEvents = CASE_TYPE.events;
-    component.selected.event = CASE_TYPE.events[1];
-    component.filterEventControl.setValue(EVENT_ID_2);
-    component.onEventIdChange();
-    const FIELD_ERRORS = [
-      {
-        message: 'This field1 failed validation'
-      },
-      {
-        message: 'This field2 failed validation'
-      }
-    ];
-    const VALID_ERROR = {
-      message: 'Field validation failed',
-      details: {
-        field_errors: FIELD_ERRORS
-      }
-    };
-    mockRouter.navigate.and.returnValue({ catch : (error) => error(VALID_ERROR)});
-
-    fixture.detectChanges();
-    let button = de.query($SELECT_BUTTON);
-    button.nativeElement.click();
-    fixture.detectChanges();
-
-    let errorElement = de.query($ERROR_SUMMARY);
-    expect(errorElement).toBeTruthy();
-    let errorMessage = errorElement.query($ERROR_MESSAGE);
-    expect(errorMessage).toBeTruthy();
-    let errorFieldMessages = errorElement.query($ERROR_FIELD_MESSAGES);
-    expect(errorFieldMessages).toBeTruthy();
-
-    component.filterJurisdictionControl.setValue(JURISDICTION_2.id);
-    component.onJurisdictionIdChange();
-    fixture.detectChanges();
-
-    errorElement = de.query($ERROR_SUMMARY);
-    expect(errorElement).toBeFalsy();
-  });
-
-  it('should remove field validation errors when errors previously but new case type selected', () => {
-    component.selected.jurisdiction = JURISDICTION_2;
-    component.selected.caseType = CASE_TYPES_2[0];
-    component.selectedJurisdictionCaseTypes = CASE_TYPES_2;
-    component.selectedCaseTypeEvents = CASE_TYPES_2[0].events;
-    component.selected.event = CASE_TYPES_2[0].events[1];
-    component.filterEventControl.setValue(EVENT_ID_2);
-    component.onEventIdChange();
-    const FIELD_ERRORS = [
-      {
-        message: 'This field1 failed validation'
-      },
-      {
-        message: 'This field2 failed validation'
-      }
-    ];
-    const VALID_ERROR = {
-      message: 'Field validation failed',
-      details: {
-        field_errors: FIELD_ERRORS
-      }
-    };
-    mockRouter.navigate.and.returnValue({ catch : (error) => error(VALID_ERROR)});
-
-    fixture.detectChanges();
-    let button = de.query($SELECT_BUTTON);
-    button.nativeElement.click();
-    fixture.detectChanges();
-
-    let errorElement = de.query($ERROR_SUMMARY);
-    expect(errorElement).toBeTruthy();
-    let errorMessage = errorElement.query($ERROR_MESSAGE);
-    expect(errorMessage).toBeTruthy();
-    let errorFieldMessages = errorElement.query($ERROR_FIELD_MESSAGES);
-    expect(errorFieldMessages).toBeTruthy();
-
-    component.filterCaseTypeControl.setValue(CASE_TYPES_2[2].id);
-    component.onCaseTypeIdChange();
-    fixture.detectChanges();
-
-    errorElement = de.query($ERROR_SUMMARY);
-    expect(errorElement).toBeFalsy();
-  });
-
-  it('should remove field validation errors when errors previously but new event selected', () => {
-    component.selected.jurisdiction = JURISDICTION_2;
-    component.selected.caseType = CASE_TYPES_2[2];
-    component.selectedCaseTypeEvents = CASE_TYPE.events;
-    component.selected.event = CASE_TYPE.events[1];
-    component.filterEventControl.setValue(EVENT_ID_2);
-    component.onEventIdChange();
-    const FIELD_ERRORS = [
-      {
-        message: 'This field1 failed validation'
-      },
-      {
-        message: 'This field2 failed validation'
-      }
-    ];
-    const VALID_ERROR = {
-      message: 'Field validation failed',
-      details: {
-        field_errors: FIELD_ERRORS
-      }
-    };
-    mockRouter.navigate.and.returnValue({ catch : (error) => error(VALID_ERROR)});
-
-    fixture.detectChanges();
-    let button = de.query($SELECT_BUTTON);
-    button.nativeElement.click();
-    fixture.detectChanges();
-
-    let errorElement = de.query($ERROR_SUMMARY);
-    expect(errorElement).toBeTruthy();
-    let errorMessage = errorElement.query($ERROR_MESSAGE);
-    expect(errorMessage).toBeTruthy();
-    let errorFieldMessages = errorElement.query($ERROR_FIELD_MESSAGES);
-    expect(errorFieldMessages).toBeTruthy();
-
-    component.filterEventControl.setValue(EVENT_ID_3);
-    component.onEventIdChange();
-    fixture.detectChanges();
-
-    errorElement = de.query($ERROR_SUMMARY);
-    expect(errorElement).toBeFalsy();
-  });
-
-  it('should change button label when callback warnings notified', () => {
-    let callbackErrorsContext: CallbackErrorsContext = new CallbackErrorsContext();
-    callbackErrorsContext.trigger_text = CreateCaseFiltersComponent.TRIGGER_TEXT_START;
-    component.callbackErrorsNotify(callbackErrorsContext);
-
-    fixture.detectChanges();
-    let button = de.query($SELECT_BUTTON);
-    expect(button.nativeElement.textContent).toEqual(CreateCaseFiltersComponent.TRIGGER_TEXT_START);
-
-    callbackErrorsContext.trigger_text = CreateCaseFiltersComponent.TRIGGER_TEXT_CONTINUE;
-    component.callbackErrorsNotify(callbackErrorsContext);
-
-    fixture.detectChanges();
-    expect(button.nativeElement.textContent).toEqual(CreateCaseFiltersComponent.TRIGGER_TEXT_CONTINUE);
-  });
-
-  it('should clear errors and warnings', () => {
-    let callbackErrorsContext: CallbackErrorsContext = new CallbackErrorsContext();
-    callbackErrorsContext.trigger_text = CaseViewerComponent.TRIGGER_TEXT_START;
-    component.callbackErrorsNotify(callbackErrorsContext);
-
-    fixture.detectChanges();
-    component.resetErrors();
-
-    let error = de.query($ERROR_SUMMARY);
-    expect(error).toBeFalsy();
-    expect(component.error).toBeFalsy();
-    expect(component.ignoreWarning).toBeFalsy();
-  });
-
-  it('should disable button when form valid but callback validation errors exists', () => {
-    component.selected.jurisdiction = JURISDICTION_2;
-    component.selected.caseType = CASE_TYPES_2[2];
-    component.selected.event = CASE_TYPE.events[1];
-    component.error = HttpError.from({});
-    fixture.detectChanges();
-
-    let button = de.query($SUBMIT_BUTTON);
-    expect(attr(button, 'disabled')).toEqual(null);
-    const FIELD_ERRORS = [
-      {
-        x: ''
-      }
-    ];
-    const VALID_ERROR = {
-      details: {
-        field_errors: FIELD_ERRORS
-      }
-    };
-    let error = HttpError.from(VALID_ERROR);
-    component.error = error;
-    fixture.detectChanges();
-
-    expect(attr(button, 'disabled')).toEqual('');
-  });
-
-  it('should disable button when form valid but callback errors exist', () => {
-    component.selected.jurisdiction = JURISDICTION_2;
-    component.selected.caseType = CASE_TYPES_2[2];
-    component.selected.event = CASE_TYPE.events[1];
-    fixture.detectChanges();
-
-    let button = de.query($SUBMIT_BUTTON);
-
-    expect(attr(button, 'disabled')).toEqual(null);
-    let error = HttpError.from({});
-    error.callbackErrors = ['anErrors'];
-    component.error = error;
-    fixture.detectChanges();
-
-    expect(attr(button, 'disabled')).toEqual('');
-  });
-
-  it('should enable button when previously disabled for callback errors but new jurisdiction, case type and event selected', () => {
-    let error = HttpError.from({});
-    error.callbackErrors = ['anErrors'];
-    component.error = error;
-    fixture.detectChanges();
-    let button = de.query($SUBMIT_BUTTON);
-    expect(attr(button, 'disabled')).toEqual('');
-
-    component.filterJurisdictionControl.setValue(JURISDICTION_2.id);
-    component.onJurisdictionIdChange();
-    component.selected.caseType = CASE_TYPES_2[2];
-    component.selected.event = CASE_TYPE.events[1];
-    component.error = null;
-    fixture.detectChanges();
-    expect(attr(button, 'disabled')).toEqual(null);
-    expect(mockCallbackErrorSubject.next).toHaveBeenCalledWith(null);
-    expect(mockAlertService.clear).toHaveBeenCalled();
-  });
 });