+++ src/shared/components/palette/number/read-number-field.component.ts
@@ -3,6 +3,6 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'ccd-read-number-field',
-  template: `<span class="text-16">{{caseField.value}}</span>`
+  template: `<span class="text-16">{{caseField.value | ccdCaseReference}}</span>`
 })
 export class ReadNumberFieldComponent extends AbstractFieldReadComponent {}