+++ src/shared/services/search/search.service.ts
@@ -1,16 +1,18 @@
 import { Injectable } from '@angular/core';
 import { RequestOptionsArgs } from '@angular/http';
-import 'rxjs/add/operator/map';
+import { map } from 'rxjs/operators';
 import { Observable } from 'rxjs';
 import { AbstractAppConfig } from '../../../app.config';
 import { HttpService } from '../http';
+import { Headers } from '@angular/http';
 import { RequestOptionsBuilder, SearchView } from '../request';
 import { SearchInput } from '../../components/search-filters';
 import { SearchResultView } from '../../domain/search';
 
 @Injectable()
 export class SearchService {
-
+  public static readonly V2_MEDIATYPE_SEARCH_INPUTS =
+    'application/vnd.uk.gov.hmcts.ccd-data-store-api.ui-search-input-details.v2+json;charset=UTF-8';
   public static readonly VIEW_SEARCH = 'SEARCH';
   public static readonly VIEW_WORKBASKET = 'WORKBASKET';
   public static readonly FIELD_PREFIX = 'case.';
@@ -32,30 +34,39 @@ export class SearchService {
 
     return this.httpService
       .get(url, options)
-      .map(response => response.json());
+      .pipe(
+        map(response => response.json())
+      );
   }
 
-  getSearchInputUrl(jurisdictionId: string, caseTypeId: string): string {
-    return `${this.appConfig.getApiUrl()}/caseworkers/:uid/jurisdictions/${jurisdictionId}/case-types/${caseTypeId}/inputs`;
+  getSearchInputUrl(caseTypeId: string): string {
+    return `${this.appConfig.getCaseDataUrl()}/internal/case-types/${caseTypeId}/search-inputs`;
   }
 
   getSearchInputs(jurisdictionId: string, caseTypeId: string): Observable<SearchInput[]> {
-    let url = this.getSearchInputUrl(jurisdictionId, caseTypeId);
+    let url = this.getSearchInputUrl(caseTypeId);
+    const headers = new Headers({
+      'Accept': SearchService.V2_MEDIATYPE_SEARCH_INPUTS,
+      'experimental': 'true',
+    });
     this.currentJurisdiction = jurisdictionId;
     this.currentCaseType = caseTypeId;
     return this.httpService
-      .get(url)
-      .map(response => {
-        let searchInputs = response.json();
-        if (this.isDataValid(jurisdictionId, caseTypeId)) {
-          searchInputs.forEach(item => {
-            item.field.label = item.label;
-          });
-        } else {
-          throw new Error('Response expired');
-        }
-        return searchInputs;
-      });
+      .get(url, { headers })
+      .pipe(
+        map(response => {
+          let jsonResponse = response.json();
+          let searchInputs = jsonResponse.searchInputs;
+          if (this.isDataValid(jurisdictionId, caseTypeId)) {
+            searchInputs.forEach(item => {
+              item.field.label = item.label;
+            });
+          } else {
+            throw new Error('Response expired');
+          }
+          return searchInputs;
+        })
+      );
   }
 
   isDataValid(jurisdictionId: string, caseTypeId: string): boolean {